RISIMP-1, RISIMP-3, RISIMP-4
This is the execution entry point for synchronizing the RIS database
to GRIIDC database. It reads the RIS database in dependancy order
and writes records into the GRIIDC tables. It collects data about
success and exceptions from called components, writes messages to
log files and email exception files to designated addresses.
This programs reads directives from /etc/griidc/db.in