jvh PELAGOS-33
Modified dif.php as follows

Changed the text in the pop up shown after successful save and submit from:
"Congratulations! You have successfully submitted a DIF to GRIIDC. THE UDI for this dataset is <UDI>.
The DIF will now be reviewed by GRIIDC staff and is locked to prevent editing.
To unlock your DIF to make changes, you must return to the DIF webpage. You can
then load the DIF form you wish to edit form the datasets list and select "Request Unlock"  from the bottom of the form."

TO:
"Congratulations! You have successfully submitted a DIF to GRIIDC. The UDI for this dataset is <UDI>.
The DIF will be reviewed and approved by GRIIDC staff and is locked to prevent editing. To make changes
to your DIF, please email GRIIDC at griidc@gomri.org with the UDI for your dataset.
Note that you will receive an email notification when your DIF is approved."