Fixed issue where projects listing was loading twice
When clicking on a project in the tree, the hash in the URL is updated.
This triggers the hash change event listener, which was reloading the
project listing. I added conditions to not select nodes in the tree that
are already selected, which fixes the issue with the project listing
loading twice. I noticed the issue because the status graph was loading
twice.