Modified entities to work with dataset UDI
Modified publication entity to use the publication DOI as its PK
Modified dataset2publication_link entity accordingly, and changed dataset_uid
to dataset_uid. This field is meant to be associated with a unique dataset
identifier, but the existing data structure does not provide a way to
uniquely identify a dataset. Therefore this is a 16 character field, and the
referential integrity will have to be enforced by the front-end application
until a proper data model can be developed.