Simplified display on dataland "Publications" tab, added doi hyperlink
Publications are displayed without a wrapper box, and the doi that was
added after the citation text has been removed as it is really not part
of the citation format, although may be included in some citations by
their creator.  Also, a hyperlink to dx.doi.org is created on each
citation.