Created tables and a view for storing/displaying person attributes.
These scripts create the base tables for person, email, and the email to
person mapping, as well as the view interface for front-end integration. The
email2person relationship table allows for many emails per person, but only
one can be identified as the primary email address. The requirements for this
task defined a single email address as an attribute of person, so the view
presents the primary email address as that attribute. The view's trigger
function, when presented with an email address, updates any existing primary
email address that is not the one presented to the funciton to not be the
primary, and the function email address parameter is then identified as the
primary address. This fulfills the task requirement while preserving future
functionality allowing multiple email addresses per person when identified
as a need.