Added person/email unique enforcement, modified error messages
The requirements call for an email address to be unique in person, so code was
added to enforce that requirement. This made the error message when such a
situation arose a bit confusing, so that was modified. Since several scenarios
now make use of the same message, the error message and error hints were
transformed into variables.