Enforced email address case insensitivity, enforced FK references properly
The underlying email2person table was still enforcing email address case
sensitivity for the address local part. That has been removed by verifying
the email address exists in a case insensitive manner.
The table also restricted deletion of person instances if they were still
referenced in the email2person table. That constraint should be CASCADE,
not RESTRICT, deleting the person instance in the email2person table as well.
It also prevented updates to the email address in the email table from
proceeding. Despite the fact that email address is the primay key in the
email table and as such should be unchanging per proper database design,
enforcing that rule in the email2person table is incorrect (because it only
enforces the rule on email addresses referenced in the email2person table,
not all email addresses in the email table). So ON UPDATE was modified from
RESTRICT to CASCADE.