Enabled ability to maintain original error coded for unknown errors.
The person view gracefully handled all errors, but unanticipated errors emitted
very little information. This change preserves the orignial SQLSTATE value and
logs it properly to aid in diagnosis, but retains the custom error message and
custom hint so front-end applications can properly trap errors based on that if
desired.