Changed included Common.php location from the old global location to new local location
The old global location is outside of any branch and is often stale in relation to
a development branch.  This changes to a local branch reference of share/php/Common.php
instead of from the branch installed in /opt/pelagos (or similar).  The relative location
was also corrected (needed one more ../).