Added person instantiation attributes
Added person_instantiation_time and person_instantiator attributes to the
person_table entity, and the appropriate attributes to the person view. Also
added checks to ensure the attributes are provided for the INSERT portion of
the trigger function, and code to handle a bad time value error gracefully.
Finally, added code to the INSERT portion of the trigger function to handle the
new attributes.
In the process, also added person_creation_time to the perosn_table to capture
record creation time.