Switched from using DOM sourced data to JavaScript sourced data. Added default sort too.
This change speeds up datatables rendering as it enables the use of a deferred
rendering option where only the page being displayed is rendered.  This
is possible for both ajax sourced data as well as javascript sourced
data, but not for DOM sourced data.  Default sort was also set on
Last Modified, descending as requested.