Modified history table and function code
Changed modification attributes to OLD attributes and added NEW attributes.
Modified view function code to make use of the new attributes.
NOTE that when performing a DELTE operation there are no NEW attributes. So a
DELTE operation will capture the time. Rather than place NULL into the NEW
modifier field, the code captures the current database user name. This is
likely to be a generic front-end user such as gomri_reader, but this will give
the ability to use defined database roles for each GRIIDC person in the future.

Also, the modification information was only being captured when person_table
attributes were UPDATEd. A change of just email address was not captured. The
function has been modified to capture all person modification information.