Set ON DELETE action to RESTRICT from CASCADE
I could swear that way back when we first started the RIS independence
epic that I asked if deletions should be restricted or cascaded, was told
cascaded, and that decision was documented within JIRA. Unfortunately trying
to navigate JIRA is such a frustratingly non-intuitive experience that it
is much easier to just make the modifications as directed to here and not
worry about trying to find previous contradictory documentation. To that end
I have modified all entity creation scripts within the RIS independence epic
to now enforce restricted deletions. This push reflects that change.

The following changes are included as well:
03_funding_organization/01_make_funding_organization_table.sql
Made several corrections to comments only.

03_funding_organization/02_make_email2funding_organization.sql
Initial funding organization was capturing NEW.modifier as the person
modifying the record. However, since this is a creation action, that field
should contian the NEW.creator value to prevent any possibility of not
capturing the actual record creator. Made that change.