Modified funding cycle name to CITEXT data type
This change is only to the view definition and function. name is cast as CITEXT
and the function correctly casts it back to TEXT for the underlying tables.

Also modified test_sql.sh, my test script, to output line numbers to aid in
determining which test was conducted when. Renamed the one example test script,
and added more tests to it.