Added jira_ticket attribute to registry table
The current system, as designed by the project system architect, does not
actually perform an UPDATE to registry entries. Instead most old information
was discarded and a new record with the few remaining attributes was created.
To compensate for this odd behavior, I created a BEFORE INSERT trigger early
on that captured an INSERT when an UPDATE was actually being performed, and
copied the old information to the new record. The addition of the jira_ticket
attribute required modification of this function. The function definition
is contained in the make_func_update_reg.sql file.

This alteration of the table also required re-defining the views that depend
on the registry table. Those definition are contained in the
make_dif_and_reg_views.sql file.