Added creation information to person entity
Modified person_table, person_history_table, and the person view to
accomodate creator and creation time instead of instantator and
instantation time.

Modified test script to output the file name being used for input so that
when it is called multiple times from a script the user can see which file
it is processing.
Modified test_data/04_fund_cycle_test.sql to generate random modifier
data.