Modified test_sql script to output SQLSTATE code for expected failures.
This way when a failing statement fails the reason for the failure can be
visually inspected to ensure the failure was for the expected reason. The next
step will be to place the expected SQLSTATE into the test file and compare that
with the actual output, which will automate test runs. This interim solution
was prompted by a failure that passed but for the wrong reason.