Changed attribute order
My coding standard is to define all primary key attributes in US English
alphabetic order, followed by FK attributes in the same order, followed by
non-key attributes in the same order. The exception to this rule is views,
which are normally intended for an end user's use so I define the attributes
in the order of imortance. This commit fixes a mis-ordering of attribute
ordering in the 01_publink/make_table_dataset2publication_link.sql and the
02_person/01_make_person_table.sql scripts.