Modification attributes are the same as creation attributes
Instead of displayin "N/A" (which didn't make very much sense upon reflection),
the columns now display the creation information which is how all other
modification attributes are populated on creation. Since UPDATEs to
person_token are disallowed there is no reason for the modification attributes
to contain anything, but the ORM requires them to be there for the view. So
nothing is being needlessly stored but the ORM is satisfied. Win-win all the
way around!