Extends metadata view to include extent_description and temporal extents.
While developing this view, it became apparent that indexes were needed on
something because front-end performance to determine which title to use was
horrific. EXPLAIN ANALYZE showed full sequential scans on the registry and
metadata tables because of the need to reference registry ids on the UDI
portion (an artifact of improper UPDATE design). Adding functional indexes on
the two tables made a significant improvement. This commit introduces those
indexes.