Renamed metadata_view attributes to be more in line with naming conventions.
The convention I have chosen is to preface the attribute name within the entity
definition as: entity_name_attribute, and the view attribute name with just the
attribute_name. I followed that convention for the metadata table, but not for
the metadata_view. This commit names the view attributes correctly. For legacy
reasons the metadata_xml attribute remains named as that.

This commit also introduces scripts that create functions to "back-populate"
the existing row's columns new attributes, and another script that calls all
of those functions individually to do so. The update script then DROPs the
no-longer needed functions as they are a one-time thing.