Change script name to reflect actual purpose.
I wrote two versions of this script, one to just modify the dataset_title
attribute, and one to modify all CHARACTER VARYING attributes. I created the
script that altered all columns first and copied it to the second. But when I
pushed the second it still had the name of the first. This commit changes the
script name and the comment in the script that references that name.