Adds metadata_file_hash attribute to registry and modifies registry trigger.
08_add_metadata_file_hash.sql adds the new column, dropping and recreating the
necessary views as well.

09_make_func_update_reg.sql modifies the existing udf_update_reg_trigger
function to carry over existing metadata_file_hash data as long as the NEW
url_metadata is the same as the OLD value, and a NEW metadata_file_hash has
not been provided (this allows front-end supplied values to override existing
values since the actual hash generation is external to the database).