Resolves a bug where hash domain types were not anchored properly.
The original constraint definitions only required a string that started with
the correct number of hex digits for the hash type, except for the
SHA256_HASH_TYPE, which was incorrectly defined to start with just 32 hex
digits. This commit introduces proper ending anchoring of the constraints, as
well as properly defining the SHA256_HASH_TYPE to require 64 hex digits.