Added connection statement to beginning of script
This script was originally written a few years ago, for what was then supposed
to be a short-lived system due to several design flaws. But this need was
realized, and this function put into version control pretty much as it
originally existed. Back then I knew I would be running the script with the
correct user privileges on the correct database, but that can no longer be
counted on. Therefore the script now starts by connecting to the correct
database as the correct user.

In other words, that was a tldr explanation of the title line. :)